Fix Statisticspage for Ubuntu Build
Fix Statisticspage for Building Block Panther GUI Wallet on Ubuntu

+++ src/qt/statisticspage.cpp
@@ -120,7 +120,7 @@ void StatisticsPage::updateStatistics()
     QString Qlpawrate = model->getLastBlockDate().toString();
 
     QString QPeers = QString::number(peers);
-    QString qVolume = QLocale(QLocale::English).toString(volume);
+    QString qVolume = QLocale(QLocale::English).toString((qlonglong)volume);
 
     if(nHeight > heightPrevious)
     {
